refactor: global component 한곳에 몰아두기
