+++ apps/web/src/components/Global/ClientProvider.tsx
@@ -5,9 +5,9 @@ import { SessionProvider } from 'next-auth/react';
 import { domMax, LazyMotion } from 'framer-motion';
 
 import QueryClientProvider from '@/apis/QueryClientProvider';
-import Monitoring from '@/components/Monitoring';
+import Monitoring from '@/components/Global/Monitoring';
 
-import SessionLoader from './SessionLoader';
+import SessionLoader from '../SessionLoader';
 
 function ClientProvider({ children }: PropsWithChildren) {
   return (
