+++ apps/web/src/components/Global/GlobalComponent.tsx
