+++ apps/web/src/components/Global/index.ts
@@ -1,0 +1,3 @@
+export { default as ClientProvider } from './ClientProvider';
+export { default as GlobalComponent } from './GlobalComponent';
+export { default as Monitoring } from './Monitoring';
