--- apps/web/src/components/GlobalComponent/index.ts
@@ -1,1 +1,0 @@
-export { default as GlobalComponent } from './GlobalComponent';
